pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clinai"
version = "0.1.0"
description = "Clinical Protocols Vector Database Pipeline"
requires-python = ">=3.9"
dependencies = ["pyyaml>=6.0", "python-frontmatter>=1.0", "pydantic>=2.0", "tqdm>=4.65"]

[tool.setuptools.packages.find]
where = ["src"]
